Started work.
Most of the actual processing will happen in tree like data structures, so
the first thing is to generate these structures automatically. This is a
(very) rudimentary implementation of a tree class generator that generates
its own AST representation (the CST is generated using an external tool,
"treecc"); a lot of shortcuts have been taken and values hardcoded.